requirements.txt
absl-py==0.10.0
adal==1.2.5
alembic==1.4.1
appdirs==1.4.4
APScheduler==3.6.3
astunparse==1.6.3
azure-common==1.1.25
azure-core==1.8.2
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==10.2.0
azure-mgmt-storage==11.2.0
azure-storage-blob==12.5.0
azureml==0.2.7
azureml-core==1.18.0
azureml-dataprep==2.4.0
azureml-dataprep-native==24.0.0
azureml-dataprep-rslex==1.2.3
azureml-mlflow==1.18.0.post1
azureml-tensorboard==1.18.0
backports.tempfile==1.0
backports.weakref==1.0.post1
Brotli==1.0.9
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
codecarbon==1.0.0
contextlib2==0.6.0
cryptography==3.2
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
databricks-cli==0.13.0
dataclasses==0.7
distro==1.5.0
docker==4.3.1
dotnetcore2==2.1.17
entrypoints==0.3
fire==0.3.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gorilla==0.3.0
grpcio==1.33.1
gunicorn==20.0.4
gviz-api==1.9.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.1
Keras-Preprocessing==1.1.2
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
mlflow==1.11.0
msal==1.5.1
msal-extensions==0.2.2
msrest==0.6.19
msrestazure==0.6.4
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.2
pathlib==1.0.1
pathspec==0.8.0
Pillow==8.0.1
plotly==4.12.0
portalocker==1.7.1
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyhocon==0.3.57
PyJWT==1.7.1
pynvml==8.0.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
querystring-parser==1.2.4
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.4.1
SecretStorage==3.1.2
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.4.1
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-profile==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
